Remove outdated cron job implementation
This change removes the existing cron job implementation, which included removing several classes related to Cron job processing and scheduling. This significant update was necessary to pave the way for a new, improved implementation that promises better efficiency and maintainability. Please note that any classes utilizing this previous cron job scheme will need to be updated.